Make the dockerfile use releases
(also do a change where we handle upload errors a bit better)

Basically I wanted to try using mix relase and so this is what we end up with.

The important parts
1) If you're gonna use alpine for running make sure you also use alpine for building
2) Ensure that you have the necessary deps installed in the image you're using for running

One annoyance with releases is that now we do a full re-compile of the whole thing (including the deps)
even if those are unchanged. This sucks, but that's mix release for you, because it cannot re-use the
stuff from mix compile (what with the output going somewhere else entirely). So using it for larger apps
with more deps might be annoying if you care about compile times.